+++ apm-sniffer/apm-sdk-plugin/lettuce-5.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/lettuce/v5/RedisCommandCompleteExceptionallyMethodInterceptor.java
@@ -38,7 +38,6 @@ public class RedisCommandCompleteExceptionallyMethodInterceptor implements Insta
             AbstractSpan span = (AbstractSpan) objInst.getSkyWalkingDynamicField();
             span.log(t);
             span.asyncFinish();
-            //help gc
             objInst.setSkyWalkingDynamicField(null);
         }
         return ret;