+++ apm-sniffer/apm-sdk-plugin/lettuce-5.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/lettuce/v5/RedisCommandCompleteMethodInterceptor.java
@@ -36,7 +36,6 @@ public class RedisCommandCompleteMethodInterceptor implements InstanceMethodsAro
         if (objInst.getSkyWalkingDynamicField() != null) {
             AbstractSpan span = (AbstractSpan) objInst.getSkyWalkingDynamicField();
             span.asyncFinish();
-            //help gc
             objInst.setSkyWalkingDynamicField(null);
         }
         return ret;